Move AvroCompatibilityChecker&Level from kafka-schema-registry to kafak-schema-registry-client
