+++ .gitignore
@@ -11,3 +11,12 @@ target/
 /docs/_build
 
 .DS_Store
+
+# Eclipse Core
+.project
+
+# JDT-specific (Eclipse Java Development Tools)
+.classpath
+
+# Eclipse Setting
+.settings
