+++ client/src/main/java/io/confluent/kafka/schemaregistry/avro/AvroCompatibilityLevel.java
